python-interview-gpt-question/requirements.txt
pandas==2.2.3
google-cloud==0.34.0